Fix close/focus URL sync and canonical open behavior
