+++ src/components/windows/CatalogSearchPrototypeWindow.tsx
@@ -6,11 +6,12 @@ import { WindowRef } from "@/components/ui/window";
 interface CatalogSearchPrototypeWindowProps {
   onClose: () => void;
   windowRef: React.RefObject<WindowRef | null>;
+  onFocus?: () => void;
 }
 
-export default function CatalogSearchPrototypeWindow({ onClose, windowRef }: CatalogSearchPrototypeWindowProps) {
+export default function CatalogSearchPrototypeWindow({ onClose, windowRef, onFocus }: CatalogSearchPrototypeWindowProps) {
   return (
-    <Window id="catalog_search_prototype" ref={windowRef} className="w-200 aspect-7/5 left-1/2 top-1/2 -translate-1/2" onClose={onClose}>
+    <Window id="catalog_search_prototype" ref={windowRef} className="w-200 aspect-7/5 left-1/2 top-1/2 -translate-1/2" onClose={onClose} onFocus={onFocus}>
       <WindowTitle>Prototype — Catalog Search</WindowTitle>
       <WindowContent className="p-0">
         <iframe
