+++ src/components/windows/PodscriberWindow.tsx
@@ -10,11 +10,13 @@ import { WindowRef } from "@/components/ui/window";
 interface PodscriberWindowProps {
   onClose: () => void;
   windowRef: React.RefObject<WindowRef | null>;
+  onFocus?: () => void;
 }
 
-function PodscriberWindow({ onClose, windowRef }: PodscriberWindowProps) {
+function PodscriberWindow({
+  onFocus, onClose, windowRef }: PodscriberWindowProps) {
   return (
-    <Window id="podscriber" ref={windowRef} className="w-96 left-1/2 top-1/2 -translate-1/2" onClose={onClose}>
+    <Window id="podscriber" ref={windowRef} className="w-96 left-1/2 top-1/2 -translate-1/2" onClose={onClose} onFocus={onFocus}>
       <WindowTitle>Podscriber</WindowTitle>
       <WindowContent>
         <video
