+++ src/components/windows/SynthesiserWindow.tsx
@@ -9,11 +9,12 @@ import { WindowRef } from "@/components/ui/window";
 interface SynthesiserWindowProps {
   onClose: () => void;
   windowRef: React.RefObject<WindowRef | null>;
+  onFocus?: () => void;
 }
 
-export default function SynthesiserWindow({ onClose, windowRef }: SynthesiserWindowProps) {
+export default function SynthesiserWindow({ onClose, windowRef, onFocus }: SynthesiserWindowProps) {
   return (
-    <Window id="synthesiser" ref={windowRef} className="w-96 left-1/2 top-1/2 -translate-1/2" onClose={onClose}>
+    <Window id="synthesiser" ref={windowRef} className="w-96 left-1/2 top-1/2 -translate-1/2" onClose={onClose} onFocus={onFocus}>
       <WindowTitle>Synthesiser</WindowTitle>
       <WindowContent>
         <p className="font-serif text-xl mb-2">Generate a Minto Pyramid synthesis of any content.</p>
